chore: create defi-referrals folder and move files
