+++ app/scripts/lib/defi-referrals/defi-referral-onchain-check.test.ts
@@ -1,5 +1,5 @@
-import { CHAIN_IDS } from '../../../shared/constants/network';
-import { GMX_REFERRAL_STORAGE_ADDRESS } from '../../../shared/constants/defi-referrals';
+import { CHAIN_IDS } from '../../../../shared/constants/network';
+import { GMX_REFERRAL_STORAGE_ADDRESS } from '../../../../shared/constants/defi-referrals';
 import { checkGmxHasReferralCode } from './defi-referral-onchain-check';
 
 const WALLET_ADDRESS = '0x1234567890abcdef1234567890abcdef12345678';
