+++ app/scripts/lib/defi-referrals/index.ts
@@ -1,0 +1,6 @@
+export {
+  createDefiReferralMiddleware,
+  ReferralTriggerType,
+  type ExtendedJSONRPCRequest,
+} from './createDefiReferralMiddleware';
+export { checkGmxHasReferralCode } from './defi-referral-onchain-check';
